+++ resources/views/bootstrap-4/components/buttons/form-button.blade.php
@@ -2,33 +2,10 @@
     {{ $attributes->merge(['class' => 'btn btn-'.$variant.($size !== null ? ' btn-'.$size : '')]) }}
     type="{{ $type }}"
     form="{!! $formId !!}"
-    @if ($title !== null)
-        data-toggle="tooltip"
-        title="{!! $title !!}"
-    @endif
-    @if ($confirm !== null)
-        data-buk-confirm="{!! $confirm !!}"
-        data-buk-confirm-modal="confirm-modal-{!! $confirmId !!}"
-    @endif
+    @include('blade-ui-kit-bootstrap::bootstrap-4.components.buttons.partials.attributes')
     @disabled($disabled)
 >
-    @if ($startContent !== null)
-        {!! $startContent !!}
-    @endif
-    @if ($slot->isEmpty())
-        @if ($hideText)
-            <span class="sr-only">
-                {!! $text !!}
-            </span>
-        @else
-            {!! $text !!}
-        @endif
-    @else
-        {!! $slot !!}
-    @endif
-    @if ($endContent !== null)
-        {!! $endContent !!}
-    @endif
+@include('blade-ui-kit-bootstrap::bootstrap-4.components.buttons.partials.content')
 </button>
 
 @push('blade-ui-kit-bs-html')
@@ -37,7 +14,3 @@
         @method($method)
     </form>
 @endpush
-
-@if ($confirm !== null)
-    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" :confirmVariant="$confirmVariant" />
-@endif