+++ resources/views/bootstrap-5/components/buttons/partials/attributes.blade.php
@@ -1,0 +1,9 @@
+@if ($title !== null)
+    data-bs-toggle="tooltip"
+    title="{!! $title !!}"
+@endif
+@if ($confirm !== null)
+    data-buk-confirm="{!! $confirm !!}"
+    data-buk-confirm-modal="confirm-modal-{!! $confirmId !!}"
+    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" :confirmVariant="$confirmVariant" />
+@endif